Moved a lot of logic out of the RenderWorld class
The RenderWorld class was getting rather bloated and adding extra renderer objects was going to get harder. Lots of the logic has been moved to the canvas class.
The canvas class has also been split into two to make it easier to follow. One stores the base logic for the canvas and the other stores the user inputs