Modified plugin system
Options in features enables the change options dialog but options can still be given to the plugin. This allows options to be edited in the select destination UI and passed to the main operation without having options before.
Also merged options and wxoptions in the inputs list to just options