Made a resource pack class
This class will handle getting the models, creating the atlas and setting up the opengl textures.
One instance of this class should be created and passed around all the objects.
This should solve the janky model solution we currently have where every instance refers back to the main class. (eg chunk back to the world)