Fixed an issue with the inspect block tool
Not exactly related to the set biome operation but I noticed this bug while implementing that.
When moving the mouse around while holding Alt the bubble would flash.
Also while moving the mouse around quickly it was possible for the bubble to show imformation from the previously highlighted block.
The solution was to check if the mouse has moved since the last frame and not show the bubble if that was the case. This solves both problems.
I also did a little reorganising while I was at it.