Fixed rotation values to match Minecraft and partially solved the transpose issue
OpenGL requires the transformation matrix to be the transpose of what it would normally be.
I made two bugs when implementing this that canceled each other out.
I am slowly pushing the transpose issue further down into the code until there will be only one transpose when handing the value over to OpenGL.
Doing it in one go lead to lots of issues that were virtually impossible to solve