Reduced the timer draw interval
The timer will run as often as it can. If the computer is not able to draw more than 30 fps it won't but if it is able to run 60 it will.
There was an issue where computers were able to manage 30 fps but the timer was not running fast enough so it was only getting about 20 fps