Switched chunk loading to a single thread rather than one per object
Previously chunk loading was done on one thread per level.
If there were multiple levels present the drawing thread would not have a chance to work and the program would become slow.
With one thread it will behave as if there was only one object but chunk loading will be spread across the objects.