Restructured the operaiton loading logic
Before all the data was stored loosely in dictionaries and controlled by loose functions.
This wrapps all that data up in classes to make everything easier to handle.
The OperationLoader classes take the export dictionary and read it as required.
The OperationManager handles a branch of operations combining the stock operations and the public operations directory.
This still needs hooking up to the actual program