Improved language support
The language system first loads en.lang then updates it with the lang file for the system's language prefix (eg de.lang) then if a lang file exists for the region then that will be loaded and merged.
This allows for multiple languages to share the same language file if they are similar enough with region modifications where required.