Fixed renderer issue in the paste tool
The renderer was modified to draw outside the working area so that users could see what they placed outside the area even if it would not be saved.
This caused an issue in the paste tool which actually extracts whole chunks and only draws the required part of them.
This should fix that.