Added a warning dialog about installing certificates on macOS
When running from source on macOS you need to run the install certificates file.
This is mentioned in the installer but a number of users has missed this and the error that is given is a little cryptic.
Hopefully this should make it more clear to those users.
Split up the lang entries to make it easier to add more errors