Added a loading bar when closing the world if it takes more than 200ms
Some worlds take a while to close (particularly Bedrock which runs the compaction logic on close) and the program will hang while this is happening.
If closing takes more than 200ms a closing dialog will be shown to let the user know that something is happening.
Fixed a recursion issue when closing from the button in the canvas.
Fixed a crash when closing from the button in the canvas and saving.