Added an error UI that also displays the traceback
A lot of users have been reporting issues and we have to ask them for the traceback because it is currently hidden in the console.
This UI displays the traceback within the UI that is shown to the user with the ability to copy the traceback to the clipboard.
This should make it easier for users to report issues.