Handled Operation UI errors and remove the UI
If an operation fails to load for some reason this will destroy the created UI and restore everything back to the starting state.
Once this is done the traceback is show to the user.
The user can modify the code to fix the issue and reload the operations and it should appear.