Cleaned up the platform and version select
These were implemented kind of badly.
The logic relied on events to propagate the changes which meant that if code changed something it would have to wait for the event loop to update before the UI and properties would update.
This fixes all of that.
The code methods update immediately and do not produce events.
Only user created changes create events.
VersionNumberChangeEvent has been removed and replaced with VersionChangeEvent
PlatformChangeEvent will only be created in the standalone PlatformSelect widget.
The state management logic is not managed by a separate group of classes which can be subclassed to implement the required methods and behaviour. UI events must change these states and the update method will push the state back to the UI.