Fixed some issues with the base identifier select
There was previously a crash if you clicked on one of the base names after updating the text. This was because the kill focus event would update the UI but a listbox event would also be generated.
The UI now updates as you type the text. This is done by detecting the char event and calling the update logic after it is processed.
EVT_TEXT is triggered by the Set method