Deferred tool creation until they are first shown
Each tool takes a non-zero amount of time to create.
Previously this was all done when the world was first loaded which slowed down loading time.
This is now done when the tool is first shown to distribute the loading and only load the tools the user wants to use.