Added an event to enable size propagation upwards
wx only propagates size changes down the UI hierarchy.
Size changes in nested elements do not propagate upwards.
This will serve as the event for UI elements to fire to trigger the canvas/containing UI to do a layout.
This stops UI elements from having to be aware of their parents.