Stopped run_operation from being called twice
Only one operation can be running at the same time otherwise some bad things may happen.
This first checks if the run_operation method is being called within itself doing some complex logic using inspect. In this case the programmer called it twice within the same operation. This adds an error to inform them of this.
It also adds a check to make sure that another operation is not already running.