Simplified the widget init method
The previous init method could be given either a state class or the translation manager and optional data from which the state would be constructed.
This was needlessly complicated.
The new init method is given just the state. That can be constructed and given or the classmethod from_data will create it and pass it to the init emulating the old behaviour.