Fixed some issues with automatic window sizing
There were a few bugs with the old code.
Fixed an issue where the window was full screen at start but showed the full screen button instead of the restore button.
The window size stored in the config file is now the minimised size. If the user wants to un-maximise the window it should restore to the previous size even after restarting the program. The old code just restored it to the full screen size.