Improved world closing UI
Previously when the world was closed there was always a dialog that popped up to notify the user that it might take a while.
This always popped up regardless of how long it took so most of the time it just flashed.
This moves the saving logic to a new thread so that it does not block the UI and the closing notification is only shown if it takes more than 0.1 seconds to close.
Also fixed an issue where the UI would become unresponsive if the world save code ran too quickly. This was due to the close notification being present but hidden.