Improved the update check mechanic
The previous method relied on custom parsing and comparison.
This switches to the packaging library which handles all of that for us.
Split up the logic a bit to make testing easier.
Replaced CheckForUpdate class with a function passed to the thread.
Versions will only be suggested if they are newer and more stable. This was the intended behaviour before but it was a little broken.