Added methods to the app to open and close worlds
Previously the mechanic to open worlds was exposed through the AmuletUI frame which meant that the top level window had to be gotten to call the methods.
Closing was implemented through a callback which was equally janky.
This exposes nice methods in the app instance to open and close levels.
The methods these call may not yet be implemented.