Improved the world page logic
Improved the disable and enable logic to be correctly handled by the event system.
Removed the custom page storage and replaced with the wx page storage. There were desync issues with maintaining a separate list.
Added veto support when changing page so that tabs can stop the change if they are still processing.