Moved operation running to a thread
This moves the actual operation to a new thread so that the UI does not lock up when an operation takes a long time to run.
This applies to function operations and generator operations but only the latter will display estimated time like before.
Also added the ability to cancel operations.
Modified the save logic to use the new operation running system.
Depreciated the old show_loading_dialog function. I don't know if third party code still uses this.