Improved the OpenGL handling
Reading into OpenGL I found some issues with how we were doing things.
OpenGL functions cannot be called until the window is shown which might be the reason some users are getting errors.
Created a function that runs the first time the window is shown. Moved all the initial OpenGL state setting into this function. Enable is run once this function is done.
Other OpenGL initialisation is moved into the drawing functions which are run after the window is shown for the first time.