Added private variable obfuscation to the compiled build
It would be useful to be able to stop third party code poking around in private variables.
Python does not have true private variables but we can get close by obfuscating the private variable names.
Python does have double underscore attributes and methods but these are not available in first party subclasses so are not very useful in a lot of cases.
True obfuscation would be difficult and require tracking variables across modules.
This implementation is much simpler and just prefixes all private variables with the same random string.