Added new simple UI api button type
Added new simple UI api button type,

The goal of this commit was to add a button type for the simple UI interface as it was missing this option by adding _create_button() which creates a button element with the button name arg as a string and the callback function you pass to it.

When you pass the function as the second arg, it will run the function that is in your plugin, the use cases include creating open directory buttons for custom directories, custom data file save locations etc.

I finally got black working and ran it on the file before making the push request.